Create _top_nav_for_home.php
I wanted to change the background color of the header part which keeps logo and menu. As I only want to make it for the home page, I have created a separate/a copy of top_nav layout for the home page.  So that I could assign a different id for home page header.